Convert Interactive Search to TypeScript (Sonarr/Sonarr@0a0e03dca)
