+++ frontend/src/Components/Table/Column.ts
@@ -1,4 +1,5 @@
 import React from 'react';
+import { SortDirection } from 'Helpers/Props/sortDirections';
 
 type PropertyFunction<T> = () => T;
 
@@ -9,6 +10,7 @@ interface Column {
   className?: string;
   columnLabel?: string;
   isSortable?: boolean;
+  fixedSortDirection?: SortDirection;
   isVisible: boolean;
   isModifiable?: boolean;
 }
