--- frontend/src/InteractiveSearch/InteractiveSearch.js
@@ -1,237 +1,0 @@
-import PropTypes from 'prop-types';
-import React, { Fragment } from 'react';
-import Alert from 'Components/Alert';
-import Icon from 'Components/Icon';
-import LoadingIndicator from 'Components/Loading/LoadingIndicator';
-import FilterMenu from 'Components/Menu/FilterMenu';
-import PageMenuButton from 'Components/Menu/PageMenuButton';
-import Table from 'Components/Table/Table';
-import TableBody from 'Components/Table/TableBody';
-import { align, icons, kinds, sortDirections } from 'Helpers/Props';
-import getErrorMessage from 'Utilities/Object/getErrorMessage';
-import translate from 'Utilities/String/translate';
-import InteractiveSearchFilterModalConnector from './InteractiveSearchFilterModalConnector';
-import InteractiveSearchRow from './InteractiveSearchRow';
-import styles from './InteractiveSearch.css';
-
-const columns = [
-  {
-    name: 'protocol',
-    label: () => translate('Source'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'age',
-    label: () => translate('Age'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'title',
-    label: () => translate('Title'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'indexer',
-    label: () => translate('Indexer'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'size',
-    label: () => translate('Size'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'peers',
-    label: () => translate('Peers'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'languageWeight',
-    label: () => translate('Languages'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'qualityWeight',
-    label: () => translate('Quality'),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'customFormatScore',
-    label: React.createElement(Icon, {
-      name: icons.SCORE,
-      title: () => translate('CustomFormatScore')
-    }),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'indexerFlags',
-    label: React.createElement(Icon, {
-      name: icons.FLAG,
-      title: () => translate('IndexerFlags')
-    }),
-    isSortable: true,
-    isVisible: true
-  },
-  {
-    name: 'rejections',
-    label: React.createElement(Icon, {
-      name: icons.DANGER,
-      title: () => translate('Rejections')
-    }),
-    isSortable: true,
-    fixedSortDirection: sortDirections.ASCENDING,
-    isVisible: true
-  },
-  {
-    name: 'releaseWeight',
-    label: React.createElement(Icon, { name: icons.DOWNLOAD }),
-    isSortable: true,
-    fixedSortDirection: sortDirections.ASCENDING,
-    isVisible: true
-  }
-];
-
-function InteractiveSearch(props) {
-  const {
-    searchPayload,
-    isFetching,
-    isPopulated,
-    error,
-    totalReleasesCount,
-    items,
-    selectedFilterKey,
-    filters,
-    customFilters,
-    sortKey,
-    sortDirection,
-    type,
-    longDateFormat,
-    timeFormat,
-    onSortPress,
-    onFilterSelect,
-    onGrabPress,
-    onForceDownloadPress
-  } = props;
-
-  const errorMessage = getErrorMessage(error);
-
-  return (
-    <div>
-      <div className={styles.filterMenuContainer}>
-        <FilterMenu
-          alignMenu={align.RIGHT}
-          selectedFilterKey={selectedFilterKey}
-          filters={filters}
-          customFilters={customFilters}
-          buttonComponent={PageMenuButton}
-          filterModalConnectorComponent={InteractiveSearchFilterModalConnector}
-          filterModalConnectorComponentProps={{ type }}
-          onFilterSelect={onFilterSelect}
-        />
-      </div>
-
-      {
-        isFetching ? <LoadingIndicator /> : null
-      }
-
-      {
-        !isFetching && error ?
-          <div>
-            {
-              errorMessage ?
-                <Fragment>
-                  {translate('InteractiveSearchResultsFailedErrorMessage', { message: errorMessage.charAt(0).toLowerCase() + errorMessage.slice(1) })}
-                </Fragment> :
-                translate('EpisodeSearchResultsLoadError')
-            }
-          </div> :
-          null
-      }
-
-      {
-        !isFetching && isPopulated && !totalReleasesCount ?
-          <Alert kind={kinds.INFO}>
-            {translate('NoResultsFound')}
-          </Alert> :
-          null
-      }
-
-      {
-        !!totalReleasesCount && isPopulated && !items.length ?
-          <Alert kind={kinds.WARNING}>
-            {translate('AllResultsAreHiddenByTheAppliedFilter')}
-          </Alert> :
-          null
-      }
-
-      {
-        isPopulated && !!items.length ?
-          <Table
-            columns={columns}
-            sortKey={sortKey}
-            sortDirection={sortDirection}
-            onSortPress={onSortPress}
-          >
-            <TableBody>
-              {
-                items.map((item) => {
-                  return (
-                    <InteractiveSearchRow
-                      key={`${item.indexerId}-${item.guid}`}
-                      {...item}
-                      searchPayload={searchPayload}
-                      longDateFormat={longDateFormat}
-                      timeFormat={timeFormat}
-                      onGrabPress={onGrabPress}
-                      onForceDownloadPress={onForceDownloadPress}
-                    />
-                  );
-                })
-              }
-            </TableBody>
-          </Table> :
-          null
-      }
-
-      {
-        totalReleasesCount !== items.length && !!items.length ?
-          <div className={styles.filteredMessage}>
-            {translate('SomeResultsAreHiddenByTheAppliedFilter')}
-          </div> :
-          null
-      }
-    </div>
-  );
-}
-
-InteractiveSearch.propTypes = {
-  searchPayload: PropTypes.object.isRequired,
-  isFetching: PropTypes.bool.isRequired,
-  isPopulated: PropTypes.bool.isRequired,
-  error: PropTypes.object,
-  totalReleasesCount: PropTypes.number.isRequired,
-  items: PropTypes.arrayOf(PropTypes.object).isRequired,
-  selectedFilterKey: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
-  filters: PropTypes.arrayOf(PropTypes.object).isRequired,
-  customFilters: PropTypes.arrayOf(PropTypes.object).isRequired,
-  sortKey: PropTypes.string,
-  sortDirection: PropTypes.string,
-  type: PropTypes.string.isRequired,
-  longDateFormat: PropTypes.string.isRequired,
-  timeFormat: PropTypes.string.isRequired,
-  onSortPress: PropTypes.func.isRequired,
-  onFilterSelect: PropTypes.func.isRequired,
-  onGrabPress: PropTypes.func.isRequired,
-  onForceDownloadPress: PropTypes.func
-};
-
-export default InteractiveSearch;
