+++ frontend/src/InteractiveSearch/InteractiveSearchType.ts
@@ -1,0 +1,3 @@
+type InteractiveSearchType = 'episode' | 'season';
+
+export default InteractiveSearchType;
