--- frontend/src/InteractiveSearch/ReleaseEpisode.ts
@@ -1,10 +1,0 @@
-interface ReleaseEpisode {
-  id: number;
-  episodeFileId: number;
-  seasonNumber: number;
-  releaseDate?: string;
-  absoluteEpisodeNumber?: number;
-  title: string;
-}
-
-export default ReleaseEpisode;
