+++ frontend/src/Series/Search/SeasonInteractiveSearchModalContent.tsx
@@ -5,7 +5,7 @@ import ModalContent from 'Components/Modal/ModalContent';
 import ModalFooter from 'Components/Modal/ModalFooter';
 import ModalHeader from 'Components/Modal/ModalHeader';
 import { scrollDirections } from 'Helpers/Props';
-import InteractiveSearchConnector from 'InteractiveSearch/InteractiveSearchConnector';
+import InteractiveSearch from 'InteractiveSearch/InteractiveSearch';
 import formatSeason from 'Season/formatSeason';
 import translate from 'Utilities/String/translate';
 
@@ -31,7 +31,7 @@ function SeasonInteractiveSearchModalContent(
       </ModalHeader>
 
       <ModalBody scrollDirection={scrollDirections.BOTH}>
-        <InteractiveSearchConnector
+        <InteractiveSearch
           type="season"
           searchPayload={{
             seriesId,
