+++ frontend/src/typings/Release.ts
@@ -1,0 +1,53 @@
+import type DownloadProtocol from 'DownloadClient/DownloadProtocol';
+import Language from 'Language/Language';
+import { QualityModel } from 'Quality/Quality';
+import CustomFormat from 'typings/CustomFormat';
+
+export interface ReleaseEpisode {
+  id: number;
+  episodeFileId: number;
+  seasonNumber: number;
+  releaseDate?: string;
+  absoluteEpisodeNumber?: number;
+  title: string;
+}
+
+interface Release {
+  guid: string;
+  protocol: DownloadProtocol;
+  age: number;
+  ageHours: number;
+  ageMinutes: number;
+  publishDate: string;
+  title: string;
+  infoUrl: string;
+  indexerId: number;
+  indexer: string;
+  size: number;
+  seeders?: number;
+  leechers?: number;
+  quality: QualityModel;
+  languages: Language[];
+  customFormats: CustomFormat[];
+  customFormatScore: number;
+  sceneMapping?: object;
+  seasonNumber?: number;
+  episodeNumbers?: number[];
+  absoluteEpisodeNumbers?: number[];
+  mappedSeriesId?: number;
+  mappedSeasonNumber?: number;
+  mappedEpisodeNumbers?: number[];
+  mappedAbsoluteEpisodeNumbers?: number[];
+  mappedEpisodeInfo: ReleaseEpisode[];
+  indexerFlags: number;
+  rejections: string[];
+  episodeRequested: boolean;
+  downloadAllowed: boolean;
+  isDaily: boolean;
+
+  isGrabbing?: boolean;
+  isGrabbed?: boolean;
+  grabError?: string;
+}
+
+export default Release;
